fix nested list
Realization: for each ogmode item, the correspoindong item is available as a prop, so no other pros need to be passed.
The only importnant thing is that on new item creation the bookkeping is done.
The levels are stored in the items, but actually the react list takes care of the indentation